feat(global-nav): set short tooltip delay for the GlobalNav & DTP Toggle (#6515)
* feat: add short tooltip delay and use it in the GlobalNav & DTP Toggle

* feat: updating the short delay to 500ms

+++ src/split-ticket-view-toggle/components/Toggle.tsx
@@ -59,7 +59,7 @@ export default function Toggle() {
                 <Tooltip
                     target={buttonId}
                     placement="right"
-                    delay={TooltipDelay.Long}
+                    delay={TooltipDelay.Short}
                 >
                     {isEnabled ? 'Expand' : 'Collapse'}
                 </Tooltip>
@@ -69,7 +69,7 @@ export default function Toggle() {
                 <Tooltip
                     target={buttonId}
                     placement="bottom-start"
-                    delay={TooltipDelay.Long}
+                    delay={TooltipDelay.Short}
                     innerProps={{
                         popperClassName: css.tooltip,
                     }}
